% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ci.R
\name{calculate_ci}
\alias{calculate_ci}
\title{calculate_ci}
\usage{
calculate_ci(distance, dbh, max_dist, type = "exponential")
}
\arguments{
\item{distance}{distance between focal tree and all others}

\item{dbh}{DBH of trees}

\item{max_dist}{maximum interaction distance}

\item{type}{kernel type to use (either "fractional", "exponential" or "epanechnikov")}
}
\value{
vector
}
\description{
Calculate competition index
}
\details{
Calculate competition index based on distances to neighbouring trees and dbh of
neighbours. Neigbours with a distance larger than \code{max_dist} are not considered.
It is used to standartize values to 0-1 for Epanechnikov kernel
}
\examples{
add examples

}
\references{
Pommerening, A., LeMay, V., Stoyan, D., 2011. Model-based analysis of the influence
of ecological processes on forest point pattern formation-A case study. Ecol. Modell. 222, 666–678.

Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135–143.
}
