% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_growth.R
\name{calculate_growth}
\alias{calculate_growth}
\title{calculate_growth}
\usage{
calculate_growth(dbh, parameters)
}
\arguments{
\item{dbh}{Numeric with DBH of target plant.}

\item{parameters}{List with all parameters.}
}
\value{
vector
}
\description{
Calculate growth
}
\details{
Calculates DBH increase (growth) of individual trees based on current DBH.
}
\references{
Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135-143.
}
