% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_random_coords.R
\name{calculate_random_coords}
\alias{calculate_random_coords}
\title{random_distance}
\usage{
calculate_random_coords(species, n = NULL, max_dist = 80,
  number_samples = 1e+06)
}
\arguments{
\item{species}{Species of tree}

\item{n}{Number of random coordinates}

\item{max_dist}{Maximum distance of seed dispersal}

\item{number_samples}{Starting number of random coordinates checked against the seed kernal function.
A higher number will decrease the speed of the function but increase the fit to the seed kernel}
}
\description{
random_distance
}
\details{
Create distances from origin depending on seed kernel function
}
