% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competition_Kernel.R
\name{Competition.Kernel}
\alias{Competition.Kernel}
\title{Competition kernel}
\usage{
Competition.Kernel(r, dbh, max_dist, type = "Epanechnikov")
}
\arguments{
\item{r}{[\code{numeric(1)}]\cr Distance between c_i and c_j}

\item{dbh}{[\code{Numeric(1)}]\cr DBH of c_j}

\item{max_dist}{[\code{numeric(1)}]\cr Maximum interaction distance.
Used to standartize values to 0-1 for Epanechnikov kernel}

\item{type}{[\code{string(1)}]\cr Kernel type to use (either "Fractional", "Exponential" or "Epanechnikov")}
}
\description{
Function to calculate competitionfor c_i experienced by c_j
}
