% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_parameters.R
\name{construct_parameters}
\alias{construct_parameters}
\title{construct_parameters}
\usage{
construct_parameters()
}
\value{
list
}
\description{
Construct parameters
}
\details{
Construct a list with all default paramerts for run_model.
}
\examples{
\dontrun{
construct_parameters()
}

}
\references{
Bilek, L., Remes, J., Zahradnik, D., 2009. Natural regeneration of senescent even-
aged beech (Fagus sylvatica L.) stands under the conditions of Central Bohemua.
Journal of Forest Science 55(4), 145-155

Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die -
partitioning tree mortality dynamics in a near-natural mixed deciduous forest.
J. Ecol. 101, 220–230.

Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology.
For. Ecol. Manage. 331, 135-143.

Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion.
Ecology, 75(6), 1794-1806.
}
