% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_ci.R
\name{deprecated_calculate_ci}
\alias{deprecated_calculate_ci}
\title{deprecated_calculate_ci}
\usage{
deprecated_calculate_ci(data, alpha, beta, max_dist)
}
\arguments{
\item{data}{matrix with x-coordinate, y-coordinate and dbh}

\item{alpha}{parameter competition kernel}

\item{beta}{parameter competition kernel}

\item{max_dist}{maximum interaction distance}
}
\value{
vector
}
\description{
Calculate competition index
}
\details{
Calculate competition index based on distances to neighbouring trees and dbh of
neighbours. Neigbours with a distance larger than \code{max_dist} are not considered.
It is used to standartize values to 0-1 for Epanechnikov kernel
}
\references{
Pommerening, A., LeMay, V., Stoyan, D., 2011. Model-based analysis of the influence
of ecological processes on forest point pattern formation-A case study. Ecol. Modell. 222, 666-678.

Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135-143.
}
