% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_distance.R
\name{deprecated_calculate_distance}
\alias{deprecated_calculate_distance}
\title{deprecated_calculate_distance}
\usage{
deprecated_calculate_distance(point_a, point_b)
}
\arguments{
\item{point_a}{coordinates of first (set of) point(s)}

\item{point_b}{coordinates of second (set of) point(s)}
}
\value{
vector
}
\description{
Calculate distance
}
\details{
Calculate the distance between two (set of) point(s). Coordinates must be a
matrix with x coordinates in the first column and y coordinates in the second
column
}
\examples{
points_a <- cbind(c(5, 1, 2, 8), c(7, 9, 2, 3))
points_b <- cbind(c(3, 4, 9, 1), c(4, 8, 1, 4))
deprecated_calculate_distance(points_a, points_b)

\dontrun{
bench::mark(deprecated_calculate_distance(points_a, points_b),
raster::pointDistance(points_a, points_b, lonlat = FALSE),
relative = TRUE, iterations = 10000)
}

}
