% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_mortality_probability.R
\name{deprecated_calculate_mortality_probability}
\alias{deprecated_calculate_mortality_probability}
\title{deprecated_calculate_mortality_probability}
\usage{
deprecated_calculate_mortality_probability(species, dbh)
}
\arguments{
\item{species}{current species}

\item{dbh}{of current tree}
}
\value{
vector
}
\description{
Calculate mortality probability
}
\details{
Calculate the mortality probability depending on the species and the current DBH.
}
\examples{
deprecated_calculate_mortality_probability(species = "Beech", dbh = 25.31)

}
\references{
Holzwarth, F., Kahl, A., Bauhus, J., & Wirth, C. (2013). Many ways to die - partitioning
tree mortality dynamics in a near-natural mixed deciduous forest. Journal of Ecology, 101(1), 220-230.
}
