% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_probability.R
\name{deprecated_calculate_probability}
\alias{deprecated_calculate_probability}
\title{deprecated_calculate_probability}
\usage{
deprecated_calculate_probability(beta, theta, max_dist)
}
\arguments{
\item{beta}{parameter dispersal kernel}

\item{theta}{parameter dispersal kernel}

\item{max_dist}{maximum interaction distance}
}
\value{
vector
}
\description{
Calculate kernel probability
}
\details{
Calculates the probability according a dispersal kernal.
}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
