% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_random_distance.R
\name{deprecated_calculate_random_distance}
\alias{deprecated_calculate_random_distance}
\title{deprecated_calculate_random_distance}
\usage{
deprecated_calculate_random_distance(species, n = NULL, max_dist = 80,
  number_samples = 1e+06)
}
\arguments{
\item{species}{Current species}

\item{n}{number of distances}

\item{max_dist}{maximum distance}

\item{number_samples}{random number used to generate distances}
}
\value{
vector
}
\description{
Calculate random coords
}
\details{
Calculates random distances following the probability function of a species-specific
seed kernel. Because a 'brute-force' approach is used, computational demanding
}
\examples{
deprecated_calculate_random_distance(species = "Beech", n = 10)

}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
