% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_calculate_seed_kernel.R
\name{deprecated_calculate_seed_kernel}
\alias{deprecated_calculate_seed_kernel}
\title{deprecated_calculate_seed_kernel}
\usage{
deprecated_calculate_seed_kernel(species, distance, max_dist)
}
\arguments{
\item{species}{Current species}

\item{distance}{distance for which probability is calculated}

\item{max_dist}{maximum distance}
}
\value{
vector
}
\description{
Calculate seed kernel
}
\details{
Calculates the probability for a certain distance a species-specific
seed kernel.
}
\examples{
deprecated_calculate_seed_kernel(species = "Beech", distance = 10, max_dist = 120)

}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
