% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input_data}
\alias{example_input_data}
\title{Example data set I}
\format{A tibble with 1,340 rows and 5 variables:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{spec}{Species of indivdual}
  \item{Class}{Type of individual}
  \item{bhd}{Diameter at breast height of individual in cm}
}}
\usage{
example_input_data
}
\description{
The data set contains an exemplary tree population within a 500x500 observation window.
Each indiviual tree is described by its x- and y-coordinate,
tree species, type diameter at breast height (DBH)
}
\keyword{datasets}
