% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth_Function_Species.R
\name{growth_function_species}
\alias{growth_function_species}
\title{Growth function}
\usage{
growth_function_species(dbh, species)
}
\arguments{
\item{dbh}{[\code{numeric(1)}]\cr DBH of target plant}

\item{species}{[\code{string(1)}]\cr Name of the species}
}
\value{
Numeric with size increase for each point
}
\description{
Internal help function that contains all growing functions for each species
}
