% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_data}
\alias{input_data}
\title{Input data}
\format{A dataframe with 14373 rows and 5 variables:

x: x coordinate

y: y coordinate

Species: species

DBH: type (e.g. "Adult")

Type: diameter at breast height}
\usage{
input_data
}
\description{
Input data  year 1999
}
\keyword{datasets}
