% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mortality_probability.R
\name{mortality_probability}
\alias{mortality_probability}
\title{Calculcate mortalitay mortality probability (not running)}
\usage{
mortality_probability(species, dbh)
}
\arguments{
\item{species}{[\code{character(1)}]\cr Character with species of tree}

\item{dbh}{[\code{Numeric(1)}]\cr DBH of tree in cm}
}
\description{
Function to model mortality of trees
}
\references{
\itemize{
\item Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die - partitioning tree mortality dynamics in a near-natural mixed deciduous forest. J. Ecol. 101, 220–230.
}
}
