% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_Seeds.R
\name{Number.Seeds}
\alias{Number.Seeds}
\title{Number of seeds}
\usage{
Number.Seeds(species, dbh)
}
\arguments{
\item{species}{[\code{string(1)}]\cr Species of tree}

\item{dbh}{[\code{numeric(1)}]\cr DBH of tree}
}
\description{
The function calculates the number of seeds for each tree depending on the DBH
}
\references{
\itemize{
\item Ribbens, E., Silander, J.A., Pacala, S.W., 1994. Seedling recruitment in forests: Calibratiing models to predict patterns of tree seedling dispersion. Ecology 75, 1794–1806.
}
}
