% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Input.R
\name{Prepare.Input}
\alias{Prepare.Input}
\title{Prepare Input}
\usage{
Prepare.Input(input, ID = NULL)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with input data}

\item{ID}{[\code{string(1)}]\cr String with the name of ID coloumn if already existing}
}
\description{
Function to prepare the input in a way that an ID coloumn exists and the data is nested
}
