% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Input.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(data, x, y, species, type, dbh)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{x}{String with name of column containing the x-coordinate.}

\item{y}{String with name of column containing the y-coordinate.}

\item{species}{String with name of column containing the species.}

\item{type}{String with name of column containing the type as string.}

\item{dbh}{String with name of column containing the DBH as dbl.}
}
\value{
tibble
}
\description{
Prepare data
}
\details{
The funcion modifies the structures of the data dataframe. This includes
renameing of the coloumns and nesting the data.
}
\examples{
\dontrun{
names(example_input_data)
prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")
}

}
