% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{prepare_input}
\usage{
prepare_input(input, x, y, species, type, dbh)
}
\arguments{
\item{input}{input dataframe}

\item{x}{name of column containing the x-coordinate}

\item{y}{name of column containing the y-coordinate}

\item{species}{name of column containing the species}

\item{type}{name of column containing the type as string}

\item{dbh}{name of column containing the DBH as dbl}
}
\value{
tibble
}
\description{
Prepare input
}
\details{
The function modifies the structures of the input dataframe. This includes renameing
of the coloumns and nesting the data.
}
\examples{
names(example_input_data)
prepare_input(input = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")

}
