% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabmp-packages.R
\docType{package}
\name{rabmp}
\alias{rabmp}
\alias{rabmp-package}
\title{rabmp: R agent-based model for plants}
\description{
Functions to model plant growth within an old-growth beech forest. All main processes (i.e. competition, 
  growth, seed dispersal, mortality) are included.
}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

Authors:
\itemize{
  \item Lena Enderle \email{lena.enderle@stud.uni-goettingen.de}
}

}
