% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_coordinates.R
\name{random_coordinates}
\alias{random_coordinates}
\title{Random Coordinates}
\usage{
random_coordinates(species, n = NULL, max_dist = 80,
  number_samples = 1e+06)
}
\arguments{
\item{species}{Species of tree}

\item{n}{Number of random coordinates}

\item{max_dist}{Maximum distance of seed dispersal}

\item{number_samples}{Starting number of random coordinates checked against the seed kernal function.
A higher number will decrease the speed of the function but increase the fit to the seed kernel}
}
\description{
Create Random coordinates depending on seed kernel function
}
