% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{read_parameters}
\usage{
read_parameters(file, return_list = TRUE, ...)
}
\arguments{
\item{file}{String with path to text file.}

\item{return_list}{Logical if true parameters are returned as list.}

\item{...}{Arguments passed on to \code{read.table}.}
}
\value{
tibble
}
\description{
Read parameters from text file
}
\details{
Construct a list with all default paramerts for run_model. The parameters include:
\itemize{
  \item First item
  \item Second item
}
}
\references{
Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die -
partitioning tree mortality dynamics in a near-natural mixed deciduous forest.
J. Ecol. 101, 220–230.


Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology.
For. Ecol. Manage. 331, 135-143.

Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion.
Ecology, 75(6), 1794-1806.
}
