% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(data, years, verbose = TRUE)
}
\arguments{
\item{data}{Input data frame.}

\item{years}{Timesteps (years) the model runs.}
}
\value{
tibble
}
\description{
Run the modeld
}
\details{
Wrapper function to run the model. Executes (i) simulate_ci (ii) simulate_growth
(iii) simulate_seed_dispersal (iv) simulate_mortality
}
\examples{
data_trees <- prepare_input(input = example_input_data, x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")

run_model(data = data_trees, years = 3)

}
