% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_abiotic.R
\name{run_model_abiotic}
\alias{run_model_abiotic}
\title{run_model_abiotic}
\usage{
run_model_abiotic(data, parameters, abiotic, plot_area = NULL, years,
  save_each = NULL, return_seedlings = FALSE, return_nested = FALSE,
  return_tibble = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Data.table with input data.}

\item{parameters}{List with all parameters..}

\item{abiotic}{RasterLayer with abiotic conditions. Should be scaled to 0 <= x <= 1.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}

\item{years}{Numeric timesteps (years) the model runs.}

\item{save_each}{Integer value specifying time step results are saved.}

\item{return_seedlings}{Logical if seeds should be included in final output.}

\item{return_nested}{Logical if TRUE the final tibble is nested.}

\item{return_tibble}{Logical if tibble should be returned}

\item{verbose}{If TRUE, prints progress report.}
}
\value{
data.table or tibble
}
\description{
Run the model
}
\details{
Wrapper function to run the model. Executes (i) simulate_ci (ii) simulate_growth
(iii) simulate_seed_dispersal (iv) simulate_mortality. The input data must be
preprocessed using \code{prepare_input}.

Parameters include ....
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")

threshold <- quantile(df_trees$dbh, probs = 0.8)

plot_area <- spatstat::owin(xrange = c(0, 500), yrange = c(0, 500))

ppp_threshold <- spatstat::ppp(x = df_trees[dbh > threshold, x],
y = df_trees[dbh > threshold, y],
window = plot_area)

hetero <- spatstat::density.ppp(ppp_threshold,  dimyx = c(250, 250))
hetero_df <- tibble::as_tibble(hetero)
hetero_ras <- raster::rasterFromXYZ(hetero_df)

parameters <- read_parameters(file = "inst/parameters.txt", sep = ";")

parameters$growth_abiotic <- 1

result <- run_model_abiotic(data = df_trees, parameters = parameters, years = 10,
abiotic = hetero_ras, plot_area = plot_area)
}

}
