% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_biotic.R
\name{run_model_biotic}
\alias{run_model_biotic}
\title{run_model_biotic}
\usage{
run_model_biotic(data, parameters, plot_area = NULL, years,
  save_each = NULL, return_seedlings = FALSE, return_nested = FALSE,
  return_tibble = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Data.table with input data.}

\item{parameters}{List with all parameters.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}

\item{years}{Numeric timesteps (years) the model runs.}

\item{save_each}{Integer value specifying time step results are saved.}

\item{return_seedlings}{Logical if seeds should be included in final output.}

\item{return_nested}{Logical if TRUE the final tibble is nested.}

\item{return_tibble}{Logical if tibble should be returned}

\item{verbose}{If TRUE, prints progress report.}
}
\value{
data.table or tibble
}
\description{
Run the model
}
\details{
Wrapper function to run the model. Executes (i) simulate_ci (ii) simulate_growth
(iii) simulate_seed_dispersal (iv) simulate_mortality. The input data must be
preprocessed using \code{prepare_input}.

Parameters include ....
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")

parameters <- read_parameters(file = "inst/parameters.txt", sep = "\\t", return_list = TRUE)

result <- run_model_biotic(data = df_trees, parameters = parameters, years = 10)
}

}
