% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed_Kernel.R
\name{Seed.Kernel}
\alias{Seed.Kernel}
\title{Seed kernel}
\usage{
Seed.Kernel(species, distance, max_dist)
}
\arguments{
\item{species}{[\code{string(1)}]\cr Species of tree}

\item{distance}{[\code{numeric(1)}]\cr Distance r from mother tree}

\item{max_dist}{[\code{numeric(1)}]\cr Maximum distance of seed dispersal}
}
\description{
The kernel function
}
\references{
\itemize{
\item Ribbens, E., Silander, J.A., Pacala, S.W., 1994. Seedling recruitment in forests: Calibratiing models to predict patterns of tree seedling dispersion. Ecology 75, 1794–1806.
}
}
