% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Competition_Index.R
\name{Simulate.Competition.Index}
\alias{Simulate.Competition.Index}
\title{Update competition index}
\usage{
Simulate.Competition.Index(input, standardized = T, max_dist = 30)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with individuals}

\item{standardized}{[\code{logical(1)}]\cr Standardize maximum CI to 1}

\item{max_dist}{[\code{numeric(1)}]\cr Maximum interaction distance between trees}
}
\description{
The function updates the competition index of the provided pattern using an Epanechnikov kernel. The competition
kernel idea is mainly based on Pommering et al. (2011) and Pommering et al. (2014).
}
