% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Seed_Dispersal.R
\name{Simulate.Seed.Dispersal}
\alias{Simulate.Seed.Dispersal}
\title{Seed dispersal}
\usage{
Simulate.Seed.Dispersal(pattern, threshold = 30, dbh = "DBH")
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr Point pattern object of the spatstat package}

\item{dbh}{[\code{String(1)}]\cr Name of the DBH mark}
}
\description{
The function simulates growth of individual points
}
\examples{
CI <- Seed.dispersal(pattern=pattern_living_trees)

}
\references{
\itemize{
\item Clark, J.S., Silman, M., Kern, R., Macklin, E., HilleRisLambers, J., 1999. Seed dispersal near and far: Patterns across temperate and tropical forests. Ecology 80, 1475–1494.
\item Millerón, M., De Heredia, U.L., Lorenzo, Z., Alonso, J., Dounavi, A., Gil, L., Nanos, N., 2013. Assessment of spatial discordance of primary and effective seed dispersal of European beech (Fagus sylvatica L.) by ecological and genetic methods. Mol. Ecol. 22, 1531–1545.
}
}
