% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{Mortality function (not running)}
\usage{
simulate_mortality(input)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr Point pattern object of the spatstat package}

\item{dbh}{[\code{String(1)}]\cr Name of the DBH mark}
}
\description{
Function to model mortality of trees
}
\references{
\itemize{
\item Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die - partitioning tree mortality dynamics in a near-natural mixed deciduous forest. J. Ecol. 101, 220–230.
}
}
