% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality_biotic.R
\name{simulate_mortality_biotic}
\alias{simulate_mortality_biotic}
\title{simulate_mortality_biotic}
\usage{
simulate_mortality_biotic(data, parameters)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}
}
\value{
data.table
}
\description{
Simulate mortality
}
\details{
Function to model mortality of trees based on logistic regression. The mortality
probability depends on the species and the DBH.
}
\references{
Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die -
partitioning tree mortality dynamics in a near-natural mixed deciduous forest.
J. Ecol. 101, 220–230.
}
