% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seed_dispersal_biotic.R
\name{simulate_seed_dispersal_biotic}
\alias{simulate_seed_dispersal_biotic}
\title{simulate_seed_dispersal_biotic}
\usage{
simulate_seed_dispersal_biotic(data, parameters, plot_area)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}
}
\value{
data.table
}
\description{
Simulate seed dispersal
}
\details{
Simulates seed dispersal by first calculating the number of seeds for each tree
and following distributing them around parental trees following a seed kernel.
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
 type = "Class", dbh = "bhd")

parameters <- read_parameters(file = "inst/parameters.txt", sep = ";")

simulate_seed_dispersal_biotic(df_trees, parameters = parameters)
}

}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.

Bilek, L., Remes, J., Zahradnik, D., 2009. Natural regeneration of senescent even-
aged beech (Fagus sylvatica L.) stands under the conditions of Central Bohemua.
Journal of Forest Science 55(4), 145-155
}
