% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_Competition_Index.R
\name{Update.Competition.Index}
\alias{Update.Competition.Index}
\title{Update competition index}
\usage{
Update.Competition.Index(input, standardized = T, max_dist = 30)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with input data}

\item{standardized}{[\code{logical(1)}]\cr Standardize maximum CI to 1}

\item{max_dist}{[\code{numeric(1)}]\cr Maximum interaction distance between trees}
}
\description{
The function updates the competition index of the provided pattern a kernel.
}
\references{
\itemize{
\item Pommerening, A., LeMay, V., Stoyan, D., 2011. Model-based analysis of the influence of ecological processes on forest point pattern formation-A case study. Ecol. Modell. 222, 666–678.
\item Pommerening, A., Maleki, K., 2014. Differences between competition kernels and traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135–143.
}
}
