% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_competition_index.R
\name{update_competition_index}
\alias{update_competition_index}
\title{update_competition_index}
\usage{
update_competition_index(input, type = "epanechnikov", max_dist = 30,
  standardized = TRUE)
}
\arguments{
\item{input}{Tibble with input data}

\item{type}{Kernel type to use (either "fractional", "exponential" or "epanechnikov")}

\item{max_dist}{Maximum interaction distance between trees}

\item{standardized}{Standardize maximum CI to 1}
}
\description{
update_competition_index
}
\details{
Update the competition index
}
\references{
\itemize{
\item Pommerening, A., LeMay, V., Stoyan, D., 2011. Model-based analysis of the influence of ecological processes on forest point pattern formation-A case study. Ecol. Modell. 222, 666–678.
\item Pommerening, A., Maleki, K., 2014. Differences between competition kernels and traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135–143.
}
}
