% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_i.R
\name{update_i}
\alias{update_i}
\title{update_i}
\usage{
update_i(data, increase = 1)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{increase}{Numeric how much i is increased.}
}
\value{
tibble
}
\description{
Update the timestep
}
\details{
Small function to increase the time step i after each simulated time step
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")

update_i(data = df_trees)
}

}
