% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_save_each.R
\name{update_save_each}
\alias{update_save_each}
\title{update_save_each}
\usage{
update_save_each(data, save_each)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{save_each}{Integer value specifying time step results are saved.}
}
\value{
tibble
}
\description{
Update the timestep
}
\details{
Small function to increase the time step i after each simulated time step
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")

update_save_each(data = df_trees)
}

}
