% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Competition_Index.R
\name{calculate_competition_index}
\alias{calculate_competition_index}
\title{calculate_competition_index}
\usage{
calculate_competition_index(distance, dbh, max_dist,
  type = "exponential")
}
\arguments{
\item{distance}{Vector of distances between focal tree and all others}

\item{dbh}{Vector with dbh in same order as distance}

\item{max_dist}{Maximum interaction distance. Used to standartize values to 0-1 for Epanechnikov kernel}

\item{type}{Kernel type to use (either "fractional", "exponential" or "epanechnikov")}
}
\description{
calculate_competition_index
}
\details{
Function to calculate competition
}
