% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Competition_Index.R
\name{calculate_competition_index}
\alias{calculate_competition_index}
\title{Calculate competition index}
\usage{
calculate_competition_index(distance, i, dbh, max_dist,
  type = "Epanechnikov")
}
\arguments{
\item{distance}{[\code{numeric(n)}]\cr Vector with distances to all neighbouring trees n}

\item{i}{[\code{numeric(1)}]\cr Counter of the focal tree}

\item{dbh}{[\code{numeric(n)}] Vector with DBHs of all neighbouring trees n. Will be indexed by i}

\item{max_dist}{[\code{numeric(1)}]\cr Maximum interaction distance between trees}

\item{type}{[\code{string(1)}]\cr Kernel type to use (either "Fractional", "Exponential" or "Epanechnikov")}
}
\description{
Internal help function to calculate the CI for a focal tree to use with purrr::map()
}
