% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_growth.R
\name{calculate_growth}
\alias{calculate_growth}
\title{calculate_growth}
\usage{
calculate_growth(dbh)
}
\arguments{
\item{dbh}{DBH of target plant}
}
\value{
vector
}
\description{
Calculate growth
}
\details{
Calculates DBH increase (growth) of individual trees based on current DBH
}
\examples{
dbh <- c(24.3, 65.2, 12.5)
calculate_growth(dbh)

}
\references{
Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135–143.
}
