% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_seeds.R
\name{calculate_seeds}
\alias{calculate_seeds}
\title{calculate_seeds}
\usage{
calculate_seeds(species, dbh)
}
\arguments{
\item{species}{String with current species.}

\item{dbh}{Numeric with DBH of current tree.}
}
\value{
numeric
}
\description{
Calculate number of seeds
}
\details{
Calculates the number of produced seeds for each tree (without reduction for
seed mortality).
}
\examples{
\dontrun{
calculate_seeds(species = "Beech", dbh = 45.25)
}

}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
