% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify_Habitat_Lambda.R
\name{Classify.Habitats.Lambda}
\alias{Classify.Habitats.Lambda}
\title{Classifiy habitats}
\usage{
Classify.Habitats.Lambda(pattern, species, kernel = "epanechnikov",
  sigma = 50, resolution = 20, n = 5, method = "jenks")
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr Point pattern of the spatstat-package}

\item{species}{[\code{Character(1)}]\cr Name of the species}

\item{kernel}{[\code{Character(1)}] Type of the kernel to use (see spatstat::density.ppp)}

\item{sigma}{[\code{Numeric(1)}] Sigma of density kernel (see spatstat::density.ppp)}

\item{resolution}{[\code{Numeric(1)}] Patch size of resulting raster (see raster::aggregate)}

\item{n}{[\code{Numeric(1)}] Number of classes (see classInt::classIntervals)}

\item{method}{[\code{Character(1)}] Method to classifiy values (see classInt::classIntervals)}

\item{resolution}{[\code{Numeric(1)}] Patch size of resulting raster (see raster::aggregate)}
}
\value{
Returns a raster of the raster-package with the classified habitats
}
\description{
Classifies habitats based on the point density
}
\examples{
habitats_beech <- Classify.Habitats.Lambda(pattern=pattern_1999, species="Beech")

}
