% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competition_Kernel.R
\name{Competition.Kernel}
\alias{Competition.Kernel}
\title{Competition kernel}
\usage{
Competition.Kernel(r, dbh, max.dist, type = "Epanechnikov")
}
\arguments{
\item{max.dist}{[\code{Numeric(1)}]\cr Maximum interaction distance. Used to standartize values to 0-1}

\item{type}{[\code{String(1)}]\cr Kernel type to use (either "Fractional", "Exponential" or "Epanechnikov")}

\item{pattern}{[\code{Numeric(1)}]\cr Distance between c_i and c_j}

\item{mark}{[\code{Numeric(1)}]\cr Size of c_j}
}
\value{
Vector with CI values for each point
}
\description{
Function to calculate competitionfor c_i experienced by c_j
}
\examples{
CI <- Competition.Index(pattern=pattern_living_trees, DBH="DBH_13", sigma=25)

}
