% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input_data}
\alias{example_input_data}
\title{Example data set}
\format{
A tibble with 1,340 rows and 5 variables:

x: x coordinate

y: y coordinate

Class: type (e.g. "Adult")

bhd: diameter at breast height
}
\usage{
example_input_data
}
\description{
The data set contains an exemplary tree population within a 500 x 500 observation
window. Each indiviual tree is described by its x- and y-coordinate, tree species,
type, and diameter at breast height (DBH)
}
\keyword{datasets}
