% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abiotic.R
\name{extract_abiotic}
\alias{extract_abiotic}
\title{extract_abiotic}
\usage{
extract_abiotic(data, abiotic)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{abiotic}{Raster with abiotic conditions.}
}
\value{
data.table
}
\description{
Update competition index
}
\details{
ADD TEXT HERE
}
\examples{
\dontrun{
ADD TEXT HERE
}

}
\references{
ADD TEXT HERE
}
