% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abiotic.R
\name{extract_abiotic}
\alias{extract_abiotic}
\title{extract_abiotic}
\usage{
extract_abiotic(data, abiotic)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{abiotic}{Raster with abiotic conditions.}
}
\value{
vector
}
\description{
Extract abiotic values
}
\details{
Function to extract value of abiotic RasterLayer at location of individuals
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")

threshold <- quantile(df_trees$dbh, probs = 0.8)

plot_area <- spatstat::owin(xrange = c(0, 500), yrange = c(0, 500))

ppp_threshold <- spatstat::ppp(x = df_trees[dbh > threshold, x],
y = df_trees[dbh > threshold, y],
window = plot_area)

hetero <- spatstat::density.ppp(ppp_threshold,  dimyx = c(250, 250))
hetero_df <- tibble::as_tibble(hetero)
hetero_ras <- raster::rasterFromXYZ(hetero_df)

extract_abiotic(data = df_trees, abiotic = hetero_ras)

}

}
