% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth_Function_Species.R
\name{Growth.Function.Species}
\alias{Growth.Function.Species}
\title{Growth function}
\usage{
Growth.Function.Species(dbh, species)
}
\arguments{
\item{dbh}{[\code{numeric(1)}]\cr DBH of target plant}

\item{species}{[\code{string(1)}]\cr Name of the species}
}
\value{
Numeric with size increase for each point
}
\description{
Internal help function that contains all growing functions for each species
}
