% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth_Function_Species.R
\name{growth_function_species}
\alias{growth_function_species}
\title{Growth function}
\usage{
growth_function_species(dbh)
}
\arguments{
\item{dbh}{DBH of target plant}
}
\value{
Numeric with size increase for each point

Chapman-Richards growth function, parameter from Pommerening: Differences between competition
kernels and traditional size-ration based competition indices used in forest ecology
}
\description{
Internal help function that contains all growing functions for each species
}
