% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_function_species.R
\name{growth_function_species}
\alias{growth_function_species}
\title{Growth function}
\usage{
growth_function_species(species, dbh)
}
\arguments{
\item{species}{Name of the species}

\item{dbh}{DBH of target plant}
}
\value{
Numeric with size increase for each point
}
\description{
Internal help function that contains all growing functions for each species
}
