% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Input.R
\name{Prepare.Input}
\alias{Prepare.Input}
\title{Prepare Input}
\usage{
Prepare.Input(input, x, y, species, type, dbh)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with input data}

\item{x}{[\code{string(1)}]\cr String with  name of column containing the x-coordinate as dbl}

\item{y}{[\code{string(1)}]\cr String with  name of column containing the y-coordinate as dbl}

\item{species}{[\code{string(1)}]\cr String with  name of column containing the species}

\item{type}{[\code{string(1)}]\cr String with  name of column containing the type as string.
\cr The types must be 'Adult', 'Seedlings' and 'Dead'}

\item{dbh}{[\code{string(1)}]\cr String with name of column containing the DBH as dbl}
}
\description{
Function to prepare the input dataframe in a tibble with the structure needed by all functions
}
