% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Prepare Input}
\usage{
prepare_input(input, x, y, species, type, dbh)
}
\arguments{
\item{input}{Tibble with input data}

\item{x}{String with  name of column containing the x-coordinate}

\item{y}{String with  name of column containing the y-coordinate}

\item{species}{String with  name of column containing the species}

\item{type}{String with  name of column containing the type as string}

\item{dbh}{String with name of column containing the DBH as dbl}
}
\description{
Function to prepare the input dataframe in a tibble with the structure needed by all functions
}
