% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabmp-packages.R
\docType{package}
\name{rabmp}
\alias{rabmp}
\alias{rabmp-package}
\title{rabmp: R agent Based Model for Plants}
\description{
Function to model plant growth based on agents
}
\author{
\strong{Maintainer}: Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)

Authors:
\itemize{
  \item Lena Enderle \email{lena.enderle@stud.uni-goettingen.de}
}

}
