% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Coordinates.R
\name{Random.Coordinates}
\alias{Random.Coordinates}
\title{Random Coordinates}
\usage{
Random.Coordinates(species, n = 100, r_max = 80, number_samples = 1e+06)
}
\description{
Create Random coordinates depending on seed kernel function
}
