% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(data, parameters, years, verbose = TRUE)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}

\item{years}{Numeric timesteps (years) the model runs.}

\item{verbose}{If TRUE, prints progress report.}
}
\value{
tibble
}
\description{
Run the model
}
\details{
Wrapper function to run the model. Executes (i) simulate_ci (ii) simulate_growth
(iii) simulate_seed_dispersal (iv) simulate_mortality. The input data must be
preprocessed using \code{prepare_input}.

Parameters include ....
}
\examples{
\dontrun{
data_trees <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")

parameters <- rabmp::read_parameters(file = "inst/parameters.txt", sep = "\\t", return_list = TRUE)

result <- run_model(data = data_trees, parameters = parameters, years = 10)
}

}
