% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_abiotic.R
\name{run_model_abiotic}
\alias{run_model_abiotic}
\title{run_model_abiotic}
\usage{
run_model_abiotic(
  data,
  parameters,
  abiotic,
  probs = c(0.05, 0.95),
  plot_area = NULL,
  years,
  save_each = NULL,
  return_seedlings = FALSE,
  return_nested = FALSE,
  return_tibble = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.table with input data.}

\item{parameters}{List with all parameters..}

\item{abiotic}{RasterLayer with abiotic conditions. Should be scaled to -1 <= x <= 1.}

\item{probs}{Quantiles used for bad and good habitat threshold.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}

\item{years}{Numeric timesteps (years) the model runs.}

\item{save_each}{Integer value specifying time step results are saved.}

\item{return_seedlings}{Logical if seeds should be included in final output.}

\item{return_nested}{Logical if TRUE the final tibble is nested.}

\item{return_tibble}{Logical if tibble should be returned}

\item{verbose}{If TRUE, prints progress report.}
}
\value{
data.table or tibble
}
\description{
Run the model
}
\details{
Wrapper function to run the model. Executes (i) simulate_ci (ii) simulate_growth
(iii) simulate_seed_dispersal (iv) simulate_mortality. The input data must be
preprocessed using \code{prepare_input}.

Parameters include ....
}
\examples{
\dontrun{
data <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", type = "Class", dbh = "bhd")

threshold <- quantile(data$dbh, probs = 0.8)

plot_area <- spatstat::owin(xrange = c(0, 500), yrange = c(0, 500))

ppp_threshold <- spatstat::ppp(x = data[dbh > threshold, x],
y = data[dbh > threshold, y], window = plot_area)

abiotic <- spatstat::density.ppp(ppp_threshold,  dimyx = c(250, 250))
abiotic <- dplyr::mutate(tibble::as_tibble(abiotic),
value = scales::rescale(value, to = c(-1, 1)))
abiotic <- raster::rasterFromXYZ(abiotic)

parameters <- read_parameters(file = "inst/parameters_abiotic.txt", sep = ";")

result <- run_model_abiotic(data = data, parameters = parameters, years = 10,
abiotic = abiotic, plot_area = plot_area)
}

}
