% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed_Kernel.R
\name{Seed.Kernel}
\alias{Seed.Kernel}
\title{Seed kernel}
\usage{
Seed.Kernel(species, r, r.max)
}
\arguments{
\item{species}{[\code{String(1)}]\cr Species of tree}

\item{r}{[\code{Numeric(1)}]\cr Distance r from mother tree}
}
\description{
The kernel function
}
\examples{
CI <- Seed.dispersal(pattern=pattern_living_trees)

}
\references{
\itemize{
\item Martínez, I., González-Taboada, F., 2009. Seed dispersal patterns in a temperate forest during a mast event: Performance of alternative dispersal kernels. Oecologia 159, 389–400.#' \item Millerón, M., De Heredia, U.L., Lorenzo, Z., Alonso, J., Dounavi, A., Gil, L., Nanos, N., 2013. Assessment of spatial discordance of primary and effective seed dispersal of European beech (Fagus sylvatica L.) by ecological and genetic methods. Mol. Ecol. 22, 1531–1545.
\item further references
}
}
