% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seed_Kernel.R
\name{Seed.Kernel}
\alias{Seed.Kernel}
\title{Seed kernel}
\usage{
Seed.Kernel(species, r, r_max)
}
\arguments{
\item{species}{[\code{String(1)}]\cr Species of tree}

\item{r}{[\code{Numeric(1)}]\cr Distance r from mother tree}
}
\description{
The kernel function
}
\examples{
CI <- Seed.dispersal(pattern=pattern_living_trees)

}
\references{
\itemize{
\item Ribbens, E., Silander, J.A., Pacala, S.W., 1994. Seedling recruitment in forests: Calibratiing models to predict patterns of tree seedling dispersion. Ecology 75, 1794–1806.
}
}
