% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Competition_Index.R
\name{Simulate.Competition.Index}
\alias{Simulate.Competition.Index}
\title{Update competition index}
\usage{
Simulate.Competition.Index(pattern, dbh, ci, standardized = T,
  max.dist = 30)
}
\arguments{
\item{pattern}{[\code{ppp(1)}]\cr Point pattern object of the spatstat package}

\item{dbh}{[\code{String(1)}]\cr Name of the weights mark}

\item{ci}{[\code{String(1)}]\cr Name of the competition index}

\item{standardized}{[\code{Logical(1)}]\cr Standardize maximum CI to 1}

\item{max.dist}{[\code{Numeric(1)}]\cr Maximum interaction distance between trees}
}
\value{
Updated point pattern
}
\description{
The function updates the competition index of the provided pattern using an Epanechnikov kernel. The competition
kernel idea is mainly based on Pommering et al. (2011) and Pommering et al. (2014).
}
\examples{
CI <- Competition.Index(pattern=pattern_living_trees, dbh="DBH_13", ci="CI")

}
