% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Growth.R
\name{Simulate.Growth}
\alias{Simulate.Growth}
\title{Update DBH}
\usage{
Simulate.Growth(input, year)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with input data}

\item{year}{[\code{Numeric(1)}]\cr Year of simulation}
}
\value{
Vector with size increase for each point
}
\description{
The function simulates growth of individual points
}
\examples{

}
