% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ci.R
\name{simulate_ci}
\alias{simulate_ci}
\title{simulate_ci}
\usage{
simulate_ci(data, parameters)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}
}
\value{
data.table
}
\description{
Update competition index
}
\details{
The function calculated a compeition index using a kernel. Competition depends on
neighbouring trees (within max_dist). If \code{standardized = TRUE}, all values
are standardized to 0 <= CI <= 1.
}
\examples{
\dontrun{
parameters <- read_parameters(file = "inst/parameters.txt", sep = "\\t")

names(example_input_data)
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")

simulate_ci(data = df_trees, parameters = parameters)
}

}
\references{
Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135-143.
}
