% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ci.R
\name{simulate_ci}
\alias{simulate_ci}
\title{simulate_ci}
\usage{
simulate_ci(input, type = "exponential", max_dist = 30,
  standardized = TRUE)
}
\arguments{
\item{input}{input dataframe}

\item{type}{kernel type (either "fractional", "exponential" or "epanechnikov")}

\item{max_dist}{maximum interaction distance between trees}

\item{standardized}{standardize maximum CI to 1}
}
\value{
tibble
}
\description{
Update competition index
}
\details{
The function calculated a compeition index using a kernel. Competition depends on
neighbouring trees (within max_dist). If \code{standardized = TRUE}, all values
are standardized to 0 <= CI <= 1.
}
\examples{
names(example_input_data)
df_tress <- prepare_input(input = example_input_data, x = "x_coord", y = "y_coord", species = "spec", type = "Class", dbh = "bhd")
simulate_ci(input = df_tress)

}
\references{
Pommerening, A., LeMay, V., Stoyan, D., 2011. Model-based analysis of the influence
of ecological processes on forest point pattern formation-A case study. Ecol. Modell. 222, 666-678.

Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135-143.
}
