% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{Update DBH}
\usage{
simulate_growth(input)
}
\arguments{
\item{input}{[\code{tibble(1)}]\cr Tibble with input data}
}
\value{
Vector with size increase for each point
}
\description{
The function simulates growth of individual points
}
