% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth_biotic.R
\name{simulate_growth_biotic}
\alias{simulate_growth_biotic}
\title{simulate_growth_biotic}
\usage{
simulate_growth_biotic(data, parameters)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}
}
\value{
tibble
}
\description{
Simulate growth
}
\details{
Simulate growth by calculating the DBH increase and add it to the current DBH.
Also, i (iteration counter) will be increase and the type re-classified based
on the DBH: trees with a DBH smaller than 10 cm get the type 'Sapling' and trees
with a DBH over 10 cm the type 'Adult'.
}
\examples{
\dontrun{
parameters <- read_parameters(file = "inst/parameters.txt", sep = "\\t")

names(example_input_data)
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")

df_trees <- simulate_ci(data = df_trees, parameters = parameters)

simulate_growth_biotic(data = df_trees, parameters = parameters)
}

}
\references{
Pommerening, A., Maleki, K., 2014. Differences between competition kernels and
traditional size-ratio based competition indices used in forest ecology. For. Ecol. Manage. 331, 135-143.
}
