% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{simulate_mortality}
\usage{
simulate_mortality(input)
}
\arguments{
\item{input}{Tibble with input data}

\item{threshold}{Trees with an DBH increase below die}

\item{time_steps}{Time frame in which DBH increase is considered}
}
\description{
simulate_mortality
}
\details{
Function to model mortality of trees
}
\references{
\itemize{
\item Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die - partitioning tree mortality dynamics in a near-natural mixed deciduous forest. J. Ecol. 101, 220–230.
}
}
