% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_Mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{simulate_mortality}
\usage{
simulate_mortality(data, parameters)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}
}
\value{
tibble
}
\description{
Simulate mortality
}
\details{
Function to model mortality of trees based on logistic regression. The mortality
probability depends on the species and the DBH.
}
\examples{
\dontrun{
names(example_input_data)
df_trees <- prepare_data(data = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")
simulate_mortality(data = df_trees)
}

}
\references{
Holzwarth, F., Kahl, A., Bauhus, J., Wirth, C., 2013. Many ways to die -
partitioning tree mortality dynamics in a near-natural mixed deciduous forest.
J. Ecol. 101, 220–230.
}
