% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seed_dispersal.R
\name{simulate_seed_dispersal}
\alias{simulate_seed_dispersal}
\title{simulate_seed_dispersal}
\usage{
simulate_seed_dispersal(input, threshold = 30)
}
\arguments{
\item{input}{Tibble with input data}

\item{threshold}{Minimum DBH threshold for reproduction}
}
\value{
vector
}
\description{
Simulate seed dispersal
}
\details{
Simulates seed dispersal by first calculating the number of seeds for each tree
and following distributing them around parental trees following a seed kernel.
}
\examples{
\dontrun{
names(example_input_data)
df_tress <- prepare_input(input = example_input_data, x = "x_coord", y = "y_coord",
species = "spec", type = "Class", dbh = "bhd")
simulate_seed_dispersal(df_trees)
}

}
\references{
Clark, J.S., Silman, M., Kern, R., Macklin, E., HilleRisLambers, J., 1999. Seed
dispersal near and far: Patterns across temperate and tropical forests. Ecology 80, 1475-1494.

Bilek 2009 Paper

Millerón, M., De Heredia, U.L., Lorenzo, Z., Alonso, J., Dounavi, A., Gil, L.,
Nanos, N., 2013. Assessment of spatial discordance of primary and effective seed
dispersal of European beech (Fagus sylvatica L.) by ecological and genetic methods. Mol. Ecol. 22, 1531-1545.

Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
