% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seed_dispersal_abiotic.R
\name{simulate_seed_dispersal_abiotic}
\alias{simulate_seed_dispersal_abiotic}
\title{simulate_seed_dispersal_abiotic}
\usage{
simulate_seed_dispersal_abiotic(data, parameters, plot_area, abiotic,
  abiotic_quantiles)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}

\item{abiotic}{RasterLayer with abiotic conditions. Should be scaled to 0 <= x <= 1.}
}
\value{
data.table
}
\description{
Simulate seed dispersal
}
\details{
Simulates seed dispersal by first calculating the number of seeds for each tree
and following distributing them around parental trees following a seed kernel.
}
\examples{
\dontrun{
df_trees <- prepare_data(data = example_input_data,
x = "x_coord", y = "y_coord", type = "Class", dbh = "bhd")

threshold <- quantile(df_trees$dbh, probs = 0.8)

plot_area <- spatstat::owin(xrange = c(0, 500), yrange = c(0, 500))

ppp_threshold <- spatstat::ppp(x = df_trees[dbh > threshold, x],
y = df_trees[dbh > threshold, y],
window = plot_area)

hetero <- spatstat::density.ppp(ppp_threshold,  dimyx = c(250, 250))
hetero_df <- tibble::as_tibble(hetero)
hetero_ras <- raster::rasterFromXYZ(hetero_df)

parameters <- read_parameters(file = "inst/parameters.txt", sep = ";")

simulate_seed_dispersal_abiotic(df_trees, parameters = parameters)
}

}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
