% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seed_dispersal_abiotic.R
\name{simulate_seed_dispersal_abiotic}
\alias{simulate_seed_dispersal_abiotic}
\title{simulate_seed_dispersal_abiotic}
\usage{
simulate_seed_dispersal_abiotic(data, parameters, plot_area, abiotic,
  abiotic_quantiles)
}
\arguments{
\item{data}{Dataframe with input data.}

\item{parameters}{List with all parameters.}

\item{plot_area}{The plot area as \code{\link{owin}} object from the \code{spatstat} package.}

\item{abiotic}{RasterLayer with abiotic conditions. Should be scaled to 0 <= x <= 1.}

\item{abiotic_quantiles}{Quantiles used for bad and good habitat threshold.}
}
\value{
data.table
}
\description{
Simulate seed dispersal
}
\details{
Simulates seed dispersal by first calculating the number of seeds for each tree
and following distributing them around parental trees following a seed kernel.
}
\references{
Ribbens, E., Silander, J. A., & Pacala, S. W. (1994). Seedling recruitment in forests:
Calibrating models to predict patterns of tree seedling dispersion. Ecology, 75(6), 1794-1806.
}
